remove migrations from module
