+++ x/globalfee/querier_test.go
@@ -2,8 +2,18 @@ package globalfee
 
 import (
 	"testing"
+	"time"
 
+	"cosmossdk.io/simapp"
+	simappparams "cosmossdk.io/simapp/params"
+	dbm "github.com/cometbft/cometbft-db"
+	"github.com/cometbft/cometbft/libs/log"
+	tmproto "github.com/cometbft/cometbft/proto/tendermint/types"
+	"github.com/cosmos/cosmos-sdk/store"
+	storetypes "github.com/cosmos/cosmos-sdk/store/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
+	paramskeeper "github.com/cosmos/cosmos-sdk/x/params/keeper"
+	paramstypes "github.com/cosmos/cosmos-sdk/x/params/types"
 	paramtypes "github.com/cosmos/cosmos-sdk/x/params/types"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
@@ -53,4 +63,26 @@ func TestQueryMinimumGasPrices(t *testing.T) {
 			assert.Equal(t, spec.expMin, gotResp.Params.MinimumGasPrices)
 		})
 	}
+}
+
+func setupTestStore(t *testing.T) (sdk.Context, simappparams.EncodingConfig, paramstypes.Subspace) {
+	t.Helper()
+	db := dbm.NewMemDB()
+	ms := store.NewCommitMultiStore(db)
+	encCfg := simapp.MakeTestEncodingConfig()
+	keyParams := sdk.NewKVStoreKey(paramstypes.StoreKey)
+	tkeyParams := sdk.NewTransientStoreKey(paramstypes.TStoreKey)
+	ms.MountStoreWithDB(keyParams, storetypes.StoreTypeIAVL, db)
+	ms.MountStoreWithDB(tkeyParams, storetypes.StoreTypeTransient, db)
+	require.NoError(t, ms.LoadLatestVersion())
+
+	paramsKeeper := paramskeeper.NewKeeper(encCfg.Marshaler, encCfg.Amino, keyParams, tkeyParams)
+
+	ctx := sdk.NewContext(ms, tmproto.Header{
+		Height: 1234567,
+		Time:   time.Date(2020, time.April, 22, 12, 0, 0, 0, time.UTC),
+	}, false, log.NewNopLogger())
+
+	subspace := paramsKeeper.Subspace(ModuleName).WithKeyTable(types.ParamKeyTable())
+	return ctx, encCfg, subspace
 }
